+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 Changelog for Pulse3D
 =====================
 
+0.30.0 (2023-01-27)
+-------------------
+
+Added:
+^^^^^^
+- Initial support for PlateMaps, including a new section in the metadata sheet and the PlateMap Labels for
+  each well on the aggregate metrics sheets
+
+
 0.29.2 (2023-01-24)
 -------------------